feat: add details component
