--- src/App.module.scss
@@ -1,15 +1,0 @@
-.app {
-  padding: 2rem;
-
-  .button {
-    border-radius: 8px;
-    border: 1px solid transparent;
-    padding: 0.6em 1.2em;
-    font-size: 1em;
-    font-weight: 500;
-    font-family: inherit;
-    background-color: #a08d8d;
-    cursor: pointer;
-    transition: border-color 0.25s;
-  }
-}
