+++ src/components/Details/Details.module.scss
@@ -1,0 +1,13 @@
+.details {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  flex-wrap: wrap;
+  align-items: flex-start;
+  align-content: space-around;
+  border: 1px solid #fff;
+  border-radius: 8px;
+  flex: 0 1 50%;
+  padding: 8px;
+  margin-top: 32px;
+}
