+++ src/components/List/Card/Card.tsx
@@ -1,4 +1,5 @@
-import type { Character } from '~/types';
+import { Link, generatePath } from 'react-router-dom';
+import { Character } from '~/types';
 import styles from './Card.module.scss';
 
 const localDate = new Intl.DateTimeFormat('en-GB', {
@@ -7,8 +8,11 @@ const localDate = new Intl.DateTimeFormat('en-GB', {
   year: 'numeric',
 });
 
-export function Card({ name, created, mass, skin_color }: Character) {
+export function Card({ name, created, mass, skin_color, url }: Character) {
   const joinedDate = localDate.format(new Date(created));
+
+  const urlObj = new URL(url);
+  const username = urlObj.pathname.split('/')[3];
 
   return (
     <div className={styles.card}>
@@ -16,6 +20,12 @@ export function Card({ name, created, mass, skin_color }: Character) {
       <span>{joinedDate}</span>
       <div>mass - {mass}</div>
       <div>skin color - {skin_color}</div>
+      <Link
+        className={styles.link}
+        to={generatePath('/search/:id', {
+          id: username,
+        })}
+      />
     </div>
   );
 }
