+++ src/components/List/List.module.scss
@@ -3,4 +3,5 @@
   flex-wrap: wrap;
   margin-top: 32px;
   gap: 16px;
+  flex: 0 1 50%;
 }
