+++ src/components/List/index.ts
@@ -1,1 +1,2 @@
 export * from './List';
+export * from './Card';
