+++ src/index.scss
@@ -7,12 +7,18 @@
 }
 
 :root {
-  font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+  --background-color: #242424;
+  --text-color: #fff;
+  --box-shadow-color: rgba(255, 255, 255, 0.7);
+  --error-color: #a08d8d;
+  --font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
+
+  font-family: var(--font-family);
 }
 
 body {
-  background-color: #242424;
-  color: #fff;
+  background-color: var(--background-color);
+  color: var(--text-color);
 }
 
 ul {
