+++ src/pages/Home/Home.module.scss
@@ -1,0 +1,33 @@
+.home {
+  padding: 2rem;
+
+  .button {
+    border-radius: 8px;
+    border: 1px solid transparent;
+    padding: 0.6em 1.2em;
+    font-size: 1em;
+    font-weight: 500;
+    font-family: inherit;
+    background-color: #a08d8d;
+    cursor: pointer;
+    transition:
+      box-shadow 0.25s,
+      filter 0.25s ease,
+      transform 0.05s;
+
+    &:hover {
+      box-shadow: 0 0 15px var(--box-shadow-color);
+      filter: brightness(1.2);
+    }
+
+    &:active {
+      transform: scale(0.98);
+    }
+  }
+
+  .wrapper {
+    display: flex;
+    gap: 16px;
+    justify-content: space-between;
+  }
+}
